Textures, Matrices
Added Perspective and Viewport Matrix calculation  into the Vertex
Shader!
It actually was not very hard, once you figure out how to translate
set-this-variable-to-this into GLSL.
One problem, pyopengl doesn't come with any sort of actual matrix
generation functionality, so that's a problem/

Added textures on top of that texture support, should hold, so long as
you save all of the images with the same bit-depth-whatever-encoding.