Many Change
Quite a lot of fixes here.

Changed from transformations to pyrr, if only so as to not have to call
.transpose() every single time.

Added a view matrix for translations, fixed the view matrix from wierd
hyperbolic mode, fixed the modelview matrix to only have rotations,
removed manual perspective.

Removed all terrain sizing from the vertex assembly, which had a side
effect of simplifying it a good bit.
Put the fill value in a constant just before the initial array
declaration, put the scaling factor into the modelview matrix.

Put a whole bunch of padding into the vertex declarations, noticed the
colours were all wrong, should be #FFFFFF, not #000000. Also, the
y-plates need four values each, not three.

Added a VAO thing. A way to group vbos together?

WASD and MMOV to change the view and modelview matrices respectively.

Also, fixed all that wierd whatever it was that was going so badly
before. It actually works now, so I must have done something right,
right?