Renovations
Moved the images to the img subfolder.

Extracted some functions and classes to their own files, the chaemera
stuff to its own, the tex/vert loading to another.

Left the glfw, shader, gl context setup, and rendering stuff together
for now, not sure if those can or should be separated? Seems like they'd
be handy to have in a shared global context.