ITLIVES
But, it still walks through walls and topples into the endless void, so
still have that to work on.

Raimv instance is given instructions on how to wallk along the left
wall, doesn't yet seem to recognise walls when facing south, but, eh.

RAIMV has a heading property, and now correctly will only rotate in a
single direction. Looks a bit odd when pacing rapidly too. Kind of
skates around on an office chair sort of thing.